requirements.txt
argparse==1.2.1
wsgiref==0.1.2
Django==1.7.1
dj-database-url==0.2.2
dj-static==0.0.5
gunicorn==18.0
psycopg2==2.5.1
static==0.4
tornado==3.1.1
Pillow
recaptcha-client
djangorestframework
ipcalc
httpagentparser
simplejson
django-storages
boto
pylibmc==1.3.0
django-pylibmc==0.5.0
django-pylibmc-sasl
pika==0.9.8
python-gcm==0.1.5
